podcast_engine/requirements.txt
python-dotenv>=1.0.0
yt-dlp>=2023.11.16
tqdm>=4.66.0
